第四次代码提交

+++ soo-server-admin/src/main/java/com/hanson/soo/admin/pojo/dto/ProductInfoDTO.java
@@ -2,12 +2,14 @@ package com.hanson.soo.admin.pojo.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductInfoDTO {
     String productId;
     String productName;
     String departure;
     String destination;
-    Float price;
+    BigDecimal price;
     Boolean status;
 }
